+++ src/pages/Dashboard.tsx
@@ -622,7 +622,7 @@ export function Dashboard() {
             <PropertiesPanel />
             <TransactionPanel />
             <EngineeringStudioPanel selectedDesign={selectedDesign} activePlan={activePlan} boq={currentBoq} onDesignOptionsGenerated={handleAiDesignOptions} onParsed={(result) => { if (result?.buildingType) setLatestBuildingType(result.buildingType) }} onTier3Plans={handleTier3Plans} onBuildingTypeChange={setSelectedBuildingType} />
-            <BoqExportPanel selectedDesign={selectedDesign} boq={currentBoq} onExport={handleExport} />
+            <BoqExportPanel selectedDesign={selectedDesign} boq={currentBoq} onExport={handleExport} activePlan={activePlan} buildingType={buildingType} />
             <EngineeringAnalysisPanel selectedDesign={selectedDesign} />
             <GovernancePanel
               selectedDesign={selectedDesign}